requirements.txt
Flask
Flask-PyMongo
Flask-Login
werkzeug
python-dotenv